requirements.txt
amqp==5.1.1
APScheduler==3.9.1.post1
asgiref==3.5.2
astroid==2.12.13
async-timeout==4.0.2
attrs==22.2.0
billiard==4.1.0
celery==5.3.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
dill==0.3.6
Django==4.1.3
django-allauth==0.51.0
django-apscheduler==0.6.2
django-crispy-forms==1.14.0
django-filter==22.1
django-lifecycle==1.0.0
django-modeltranslation==0.18.7
djangorestframework==3.14.0
dnspython==2.2.1
drf-yasg==1.21.5
eventlet==0.33.2
exceptiongroup==1.1.0
flake8==6.0.0
greenlet==2.0.1
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.2
lazy-object-proxy==1.9.0
Markdown==3.4.1
MarkupSafe==2.1.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==22.0
platformdirs==2.5.4
pluggy==1.0.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pylint==2.15.6
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
redis==4.6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
signals==0.0.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
urlman==2.0.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
